refactor(admin): split dashboard sections
